% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Indices.R
\name{ARPB_All}
\alias{ARPB_All}
\title{ARPB_All}
\usage{
ARPB_All(Lambda, UniLambda, standardized = TRUE)
}
\arguments{
\item{Lambda}{is a matrix of factor loadings or an object that can be converted to a matrix of factor loadings by \code{\link{getLambda}}}

\item{UniLambda}{is a matrix of factor loadings or an object that can be converted to a matrix of factor loadings by \code{\link{getLambda}}}

\item{standardized}{}
}
\value{
a list where the first element is a vector of absolute relative bias by item, and the second
item is
}
\description{
\code{ARPB} computes absolute relative bias in factor loadings between the
general factor of a bifactor model and a unidimensional model.
}
\examples{

\dontrun{
ARPB(MplusAutomation::readModels(file.choose), MplusAutomation::readModels(file.choose))
}


}
\seealso{
\code{\link{bifactorIndices}}
}
