% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploratory_BifactorIndices.R
\name{bifactorIndices_EFA}
\alias{bifactorIndices_EFA}
\title{bifactorIndices_EFA}
\usage{
bifactorIndices_EFA(Lambda, ItemsBySF = NULL, LoadMin = 0.2)
}
\arguments{
\item{Lambda}{is a factor loading matrix from EFA or an object which can be converted to such.
Currently only \code{psych} objects are supported.}

\item{ItemsBySF}{is a list, indexed by factor, of vectors of item names belonging to each
factor. You must include the general factor in this list, and the list must have names which
match the factor names in \code{Lambda}. It is recommended you look at the EFA solution first
to see which factor is which. Defaults to \code{NULL}, in which case composition of specific
factors in automated by comparing loadings to \code{LoadMin}}

\item{LoadMin}{is the minimum loading size so that an item is considered to "belong" to a factor.
If \code{ItemsBySF} is not provided, then items are assigned to factors based on whether their
loading on that factor is greater than \code{LoadMin}. If \code{ItemsBySF} is provided, then
warnings are issued whenever items load above \code{LoadMin} on factors to which they do not belong,
or do not load above \code{LoadMin} on factors to which they do belong,}
}
\value{
A list of bifactor indices, including three different ECV indices, Omega, and
OmegaH.
}
\description{
Computes all available bifactor indices for the input given.
}
\details{
Only standardized models are considered for exploratory models. PUC and ARPB are not
supported for exploratory models currently, although that may change.
}
\seealso{
\code{\link{bifactorIndices}},
         \code{\link{bifactorIndicesMplus}},
         \code{\link{bifactorIndicesMplus_EFA}},
         \code{\link{bifactorIndices_EFA}},
         \code{\link{ECV_SS}},
         \code{\link{ECV_SG}},
         \code{\link{ECV_GS}},
         \code{\link{Omega_S}},
         \code{\link{Omega_H}},
}
