% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BifactorIndicesCalculator.R
\docType{package}
\name{BifactorIndicesCalculator}
\alias{BifactorIndicesCalculator}
\alias{BifactorIndicesCalculator-package}
\title{BifactorIndicesCalculator: A package for computing statistical indices relevant to bifactor measurement models.}
\description{
The BifactorIndicesCalculator package provides functions use to compute
indices described in Rodriguez et al. (2016).
}
\section{dmacs functions}{

The BifactorIndicesCalculator package includes helper functions that can compute the various
indices from Rodriguez et al. (2016) for fitted lavaan or mirt objects
(\code{\link{bifactorIndices}}) and Mplus output files
(\code{\link{bifactorIndicesMplus}}). For users of other software and those
interested, a matrix of standardized factor loadings can also be input to
(\code{\link{bifactorIndices}}).
}

\section{References}{

Rodriguez, A., Reise, S. P., & Haviland, M. G. (2016).
Evaluating bifactor models: calculating and interpreting
statistical indices. \emph{Psychological Methods, 21}(2), 137.
}

