% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BifactorIndicesCalculator.R
\name{bifactorIndicesMPlus}
\alias{bifactorIndicesMPlus}
\title{bifactorIndicesMPlus}
\usage{
bifactorIndicesMPlus(Lambda = file.choose(), UniLambda = NA,
  standardized = TRUE)
}
\arguments{
\item{Lambda}{is an Mplus output file. Defaults to an open file dialog box}

\item{UniLambda}{is an object that the function can convert to a matrix of
factor loadings. See examples. Defaults to NA}

\item{standardized}{lets the function know whether it should be looking in
the unstandardized results or the STDYX results from the Mplus output.}
}
\value{
a list of bifactor indices
}
\description{
Computes all available bifactor indices for
}
\examples{
bifactorIndicesMPlus()


UniLambda <- MPlusAutomation::readModels(file.choose())
bifactorIndicesMPlus(UniLambda = UniLambda, standardized = FALSE)

}
\seealso{
\code{\link{bifactorIndices}}
}
