% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECV_Indices.R
\name{ECV_SS_All}
\alias{ECV_SS_All}
\title{ECV_SS_All}
\usage{
ECV_SS_All(Lambda, standardized = TRUE)
}
\arguments{
\item{Lambda}{is a matrix of factor loadings or an object the function can convert to a matrix of factor loadings}

\item{standardized}{if the factor loading matrix needs to be extracted, \code{standardized} tells the function whether to extract standardized or unstandardized loadings}
}
\value{
A vector of ECVs for all factors
}
\description{
Computes ECV index for a all factors. Here, ECV is compute with respect to items which load on the factor. In the Excel version of the bifactor indices calculator, this index is referred to as ECV (NEW)
}
\seealso{
\code{\link{ECV_SS}}, \code{\link{bifactorIndices}}, \code{\link{bifactorIndicesMPlus}}
}
