% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECV_Indices.R
\name{ECV_SS_C}
\alias{ECV_SS_C}
\title{ECV_SS_C}
\usage{
ECV_SS_C(Fac, Lambda, standardized = TRUE)
}
\arguments{
\item{Fac}{is the factor for which ECV is being requested}

\item{Lambda}{is a matrix of factor loadings or an object the function can convert to a matrix of factor loadings}

\item{standardized}{if the factor loading matrix needs to be extracted, \code{standardized} tells the function whether to extract standardized or unstandardized loadings}
}
\value{
A \code{numeric}, the ECV of the factor with respect to the items in that factor
}
\description{
Computes ECV index for a single factor. Here, ECV is compute with respect to items which load on the factor. In the Excel version of the bifactor indices calculator, this index is referred to as ECV (NEW)
}
\seealso{
\code{\link{ECV_SS_All}}
}
