% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getGen}
\alias{getGen}
\title{getGen}
\usage{
getGen(Lambda)
}
\arguments{
\item{Lambda}{is a factor loading matrix}
}
\value{
The index of the general factor, or \code{NULL} if there is no general factor
}
\description{
\code{getGen} detects whether or not a single factor loads on all items, and returns the column
index of the general factor if it exists.
}
