% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getTheta}
\alias{getTheta}
\title{getTheta}
\usage{
getTheta(x, standardized = TRUE, ...)
}
\arguments{
\item{x}{asdgfa}

\item{standardized}{If}

\item{...}{Currently, no other options are available}
}
\value{
a vector of residual variance for items
}
\description{
getTheta computes or extracts a vector of residual variance for items. If a factor loading matrix is provided, then the
vector of residual variances is computed from that matrix.
If x is a fitted model and standardized residual variances are requested then a matrix of standardized factor loadings
is computed, and the vector of residual variances is computed from that matrix.
If x is a fitted model and unstandardized residual variances are requested, then the residual variances are extracted
from the fitted model. Lavaan, mirt (SingleGroupClass), and Mplus (mplus.model) models are supported.
}
\seealso{
\code{\link{getLambda}}
}
