% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{isBifactor}
\alias{isBifactor}
\title{isBifactor}
\usage{
isBifactor(Lambda)
}
\arguments{
\item{Lambda}{Matrix of factor loadings}
}
\value{
Logical. If each item loads on a general factor and at most one specific factor, returns TRUE. Otherwise FALSE.
}
\description{
Determines whether a model has bifactor structure
}
