% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MTMM_data}
\alias{MTMM_data}
\title{Simulated data for multi-trait multi-method}
\format{A data frame with 500 rows and 27 columns
The item names indicate the trait number, method number, and item number}
\usage{
MTMM_data
}
\description{
A dataset containing 500 responses to a hypothetical 27 item instrument
measuring three traits from three different sources. Each trait is measured
by 3 items collected from each of three sources (methods).
}
\keyword{datasets}
