% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Indices.R
\name{PUC}
\alias{PUC}
\title{PUC}
\usage{
PUC(Lambda, standardized = TRUE)
}
\arguments{
\item{Lambda}{is a matrix of factor loadings or an object that can be converted to a matrix of factor loadings by \code{\link{getLambda}}}
}
\value{
\code{numeric}
}
\description{
\code{PUC} computes the proportion of uncontaminated correlations for a bifactor mode
}
\examples{

\dontrun{
ARPB(MplusAutomation::readModels(file.choose), MplusAutomation::readModels(file.choose))
}

}
\seealso{
\code{\link{bifactorIndices}}
}
