% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRS_data}
\alias{SRS_data}
\title{Response Data to the SRS-22r}
\format{A data frame with 500 rows and 20 columns
Rownames indicate the item number}
\usage{
SRS_data
}
\description{
A dataset containing 500 responses to the 20 item SRS-22r scoliosis
quality of life outcome measure.
}
\details{
The Function subscale consists of items 5, 9, 12, 15, and 18.
The Pain subscale consists of items 1, 2, 8, 11, and 17.
The SelfImage subscale consists of items 4, 6, 10, 14, and 19.
The MentalHealth subscale consists of items 3, 7, 13, 16, and 20.
}
\keyword{datasets}
