% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BifactorIndicesCalculator.R
\name{bifactorIndices}
\alias{bifactorIndices}
\title{bifactorIndices}
\usage{
bifactorIndices(Lambda, Theta = getTheta(Lambda), UniLambda = NULL,
  standardized = TRUE)
}
\arguments{
\item{Lambda}{is a matrix of factor loadings or an object that can be converted to a matrix of factor loadings by \code{\link{getLambda}}.}

\item{Theta}{is a vector of residual variances. If omitted, Theta will be computed from input for Lambda}

\item{UniLambda}{is a matrix of factor loadings or an object that can be converted to a matrix of factor loadings by \code{\link{getLambda}}}

\item{standardized}{lets the function know whether to look for standardized or unstandardized results from Mplus or lavaan. if lambda is not an Mplus or lavaan object, then \code{standardized} will be ignored}
}
\value{
a list of bifactor indices
}
\description{
Computes all available bifactor indices for the input given
}
\seealso{
\code{\link{bifactorIndicesMplus}},
         \code{\link{ECV_SS_All}},
         \code{\link{ECV_SG_All}},
         \code{\link{ECV_GS_All}},
         \code{\link{IECV_All}},
         \code{\link{Omega_S_All}},
         \code{\link{Omega_H_All}},
         \code{\link{ARPB_All}}
}
