% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECV_Indices.R
\name{ECV_SG}
\alias{ECV_SG}
\title{ECV_SG}
\usage{
ECV_SG(Lambda)
}
\arguments{
\item{Lambda}{is a matrix of factor loadings. Be sure that all factors have the same variance before calling this function.}
}
\value{
A vector of ECVs for all factors
}
\description{
Computes an ECV index for all factors which can be interpreted as the proportion of common variance of the items in each factor which is due to the general factor. Here, ECV is computed only with respect to all items using the specific factor loadings in the numerator. Neither Stucky and Edelen (2015) or the Excel version of the Bifactor Indices Calculator (Dueber, 2017) use this form of ECV.
}
\details{
\code{ECV_SG} is called by \code{\link{bifactorIndices}} and \code{\link{bifactorIndicesMplus}}, which are the only functions in this package intended for casual users
}
\section{References}{

Dueber, D. M. (2017). Bifactor Indices Calculator: A Microsoft Excel-based tool to calculate various indices relevant to bifactor CFA models. <doi:10.13023/edp.tool.01>
Stucky, B. D., & Edelen, M. O. (2015). Using hierarchical IRT models to create unidimensional measures from multidimensional data. In S. P. Reise & D. A. Revicki (Eds.), Handbook of item response theory modeling: Applications to typical performance assessment (pp.183-206). New York: Routledge.
}

\examples{
Lambda <- matrix(c(.82, .10,   0,   0,
                   .77, .35,   0,   0,
                   .79, .32,   0,   0,
                   .66, .39,   0,   0,
                   .51,   0, .71,   0,
                   .56,   0, .43,   0,
                   .68,   0, .13,   0,
                   .60,   0, .50,   0,
                   .83,   0,   0, .47,
                   .60,   0,   0, .27,
                   .78,   0,   0, .28,
                   .55,   0,   0, .75),
                   ncol = 4, byrow = TRUE)
colnames(Lambda) <- c("General", "PS", "HA", "SA")
ECV_SG(Lambda)

}
\seealso{
\code{\link{ECV_SS}}, \code{\link{ECV_GS}}, \code{\link{bifactorIndices}}, \code{\link{bifactorIndicesMplus}}
}
