% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getLambda}
\alias{getLambda}
\title{getLambda}
\usage{
getLambda(x, ...)
}
\arguments{
\item{x}{an object to be converted into a factor loading matrix, or an object containing a fitted model from which
a factor loading matrix will be extracted. Supported classes are data.frame, mplus.model, lavaan, and SingleGroupClass}

\item{...}{can be used to specify whether a standardized or unstandardized factor loading matrix
should be returned. Only relevant for lavaan and mplus.model input. The standardized matrix for mplus.model
is taken from stdyx results.}
}
\value{
a matrix of factor loadings
}
\description{
getLambda computes or extracts a matrix of factor loadings given some input. Methods exist to support an input of
a dataframe, an mplus.model from MplusAutomation, a SingleGroupClass object from mirt, and a
lavaan object from lavaan. Please do not use tibbles, as they do not support row names, and it is
best if your items are given names.
}
